Refactor move riskAnalysisTemplate definitions to module risk-analysis-template folder
