+++ packages/models/src/purpose-template/protobufConverterToV2.ts
@@ -10,9 +10,9 @@ import {
   RiskAnalysisFormTemplate,
   RiskAnalysisTemplateAnswerAnnotation,
   RiskAnalysisTemplateAnswerAnnotationDocument,
-} from "../risk-analysis/riskAnalysisTemplate.js";
-import { dateToBigInt } from "../utils.js";
+} from "../risk-analysis-template/riskAnalysisTemplate.js";
 import { toTenantKindV2 } from "../tenant/protobufConverterToV2.js";
+import { dateToBigInt } from "../utils.js";
 import {
   PurposeTemplate,
   purposeTemplateState,
